Moved algorithm to seperate file. Updated README
