+++ flaskr/templates/blog/create.html
@@ -6,7 +6,7 @@
 
 {% block content %}
   <form method="post" enctype=multipart/form-data>
-    <label for="title">Title</label>
+    <label for="title">Trip Name</label>
     <input name="title" id="title" value="{{ request.form['title'] }}" required>
     <input type=file name=file required>
     <input type=submit value=Upload>
